# cj-platform-server/openapi-check.json
{"openapi":"3.1.0","info":{"title":"创界新后台 API","description":"admin-api OpenAPI. Login then Authorize with accessToken (Bearer JWT).","version":"0.1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"岗位管理","description":"岗位列表与增改"},{"name":"角色管理","description":"角色列表与增改"},{"name":"认证","description":"登录、登出、刷新 Token、当前用户"},{"name":"系统探活","description":"S0 脚手架探活，非业务"},{"name":"用户管理","description":"用户分页/增改/启停"},{"name":"部门管理","description":"部门树与增改"}],"paths":{"/admin-api/v1/platform/users/{id}":{"get":{"tags":["用户管理"],"operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringObject"}}}}}},"put":{"tags":["用户管理"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/admin-api/v1/platform/roles/{id}":{"put":{"tags":["角色管理"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/admin-api/v1/platform/posts/{id}":{"put":{"tags":["岗位管理"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/admin-api/v1/platform/depts/{id}":{"put":{"tags":["部门管理"],"operationId":"update_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeptSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/admin-api/v1/platform/users":{"get":{"tags":["用户管理"],"summary":"用户分页","operationId":"page","parameters":[{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":1}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":20}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringObject"}}}}}},"post":{"tags":["用户管理"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringLong"}}}}}}},"/admin-api/v1/platform/users/{id}/enable":{"post":{"tags":["用户管理"],"operationId":"enable","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/admin-api/v1/platform/users/{id}/disable":{"post":{"tags":["用户管理"],"operationId":"disable","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/admin-api/v1/platform/roles":{"get":{"tags":["角色管理"],"summary":"角色列表","operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPlatformRole"}}}}}},"post":{"tags":["角色管理"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringLong"}}}}}}},"/admin-api/v1/platform/posts":{"get":{"tags":["岗位管理"],"summary":"岗位列表","operationId":"list_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPlatformPost"}}}}}},"post":{"tags":["岗位管理"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringLong"}}}}}}},"/admin-api/v1/platform/depts":{"post":{"tags":["部门管理"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeptSaveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringLong"}}}}}}},"/admin-api/v1/auth/refresh":{"post":{"tags":["认证"],"summary":"刷新 Token","description":"用 refreshToken 换发新的 access/refresh","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTokenResponse"}}}}},"security":[]}},"/admin-api/v1/auth/logout":{"post":{"tags":["认证"],"summary":"登出","description":"可选传 refreshToken 作废服务端刷新令牌","operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}},"security":[]}},"/admin-api/v1/auth/login":{"post":{"tags":["认证"],"summary":"登录","description":"账号密码登录；成功返回 accessToken 与 refreshToken","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTokenResponse"}}}}},"security":[]}},"/admin-api/v1/system/ping":{"get":{"tags":["系统探活"],"summary":"Ping","operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringObject"}}}}},"security":[]}},"/admin-api/v1/platform/depts/tree":{"get":{"tags":["部门管理"],"summary":"部门树","operationId":"tree","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListMapStringObject"}}}}}}},"/admin-api/v1/auth/me":{"get":{"tags":["认证"],"summary":"当前用户","description":"需 Authorization: Bearer {accessToken}","operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserSummary"}}}}}}}},"components":{"schemas":{"ApiResponseVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{},"message":{"type":"string"},"requestId":{"type":"string"}}},"UserSaveRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string"},"displayName":{"type":"string","minLength":1},"mobile":{"type":"string"},"email":{"type":"string"},"userType":{"type":"string"},"deptId":{"type":"integer","format":"int64"},"roleIds":{"type":"array","items":{"type":"integer","format":"int64"}}},"required":["displayName","username"]},"RoleSaveRequest":{"type":"object","properties":{"code":{"type":"string","minLength":1},"name":{"type":"string","minLength":1},"remark":{"type":"string"},"dataScope":{"type":"string"}},"required":["code","name"]},"PostSaveRequest":{"type":"object","properties":{"code":{"type":"string","minLength":1},"name":{"type":"string","minLength":1},"sortNo":{"type":"integer","format":"int32"}},"required":["code","name"]},"DeptSaveRequest":{"type":"object","properties":{"parentId":{"type":"integer","format":"int64"},"name":{"type":"string","minLength":1},"sortNo":{"type":"integer","format":"int32"}},"required":["name"]},"ApiResponseMapStringLong":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}},"message":{"type":"string"},"requestId":{"type":"string"}}},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"ApiResponseTokenResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/TokenResponse"},"message":{"type":"string"},"requestId":{"type":"string"}}},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserSummary"}}},"UserSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"displayName":{"type":"string"},"userType":{"type":"string"},"deptId":{"type":"integer","format":"int64"},"roleCodes":{"type":"array","items":{"type":"string"}},"permissions":{"type":"array","items":{"type":"string"}},"dataScope":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","username"]},"ApiResponseMapStringObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object","additionalProperties":{}},"message":{"type":"string"},"requestId":{"type":"string"}}},"ApiResponseListPlatformRole":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PlatformRole"}},"message":{"type":"string"},"requestId":{"type":"string"}}},"PlatformRole":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"remark":{"type":"string"},"status":{"type":"string"},"dataScope":{"type":"string"},"creator":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"updater":{"type":"string"},"updateTime":{"type":"string","format":"date-time"},"deleted":{"type":"boolean"},"version":{"type":"integer","format":"int32"}}},"ApiResponseListPlatformPost":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PlatformPost"}},"message":{"type":"string"},"requestId":{"type":"string"}}},"PlatformPost":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"sortNo":{"type":"integer","format":"int32"},"status":{"type":"string"},"creator":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"updater":{"type":"string"},"updateTime":{"type":"string","format":"date-time"},"deleted":{"type":"boolean"},"version":{"type":"integer","format":"int32"}}},"ApiResponseListMapStringObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"type":"object","additionalProperties":{}}},"message":{"type":"string"},"requestId":{"type":"string"}}},"ApiResponseUserSummary":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/UserSummary"},"message":{"type":"string"},"requestId":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"accessToken from /admin-api/v1/auth/login","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}